requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.1.0.dev0
anyio==3.6.0
apturl==0.5.2
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.5.7
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==5.0.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.9
class-registry==2.1.2
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
discord-components==2.1.2
discord.py==1.7.3
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-filter==21.1
django-picklefield==3.0.1
django-recaptcha==3.0.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
duplicity==0.8.12.0
email-validator==1.2.1
entrypoints==0.3
fastapi==0.77.1
fasteners==0.14.1
filelock==3.6.0
filters==1.3.2
filters-django==1.0.5
frozenlist==1.3.0
future==0.18.2
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
gunicorn==20.1.0
h11==0.9.0
httplib2==0.14.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
monotonic==1.5
multidict==6.0.2
netifaces==0.10.4
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
orjson==3.6.8
paramiko==2.6.0
pbr==5.9.0
pdfkit==1.0.0
pexpect==4.6.0
Pillow==9.0.0
pipenv==2022.5.2
platformdirs==2.5.2
protobuf==3.20.1
psutil==5.9.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.7.0
pycups==1.9.73
pydantic==1.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.4
python-distutils-extra==2.39
python-dotenv==0.20.0
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
regex==2022.6.2
requests==2.27.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
spotipy==2.19.0
SQLAlchemy==1.3.12
sqlparse==0.4.2
starlette==0.19.1
systemd-python==234
testresources==2.0.1
toml==0.10.2
typing_extensions==4.2.0
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.2.0
unattended-upgrades==0.1
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.7
usb-creator==0.3.7
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
wadllib==1.3.3
waitress==2.0.0
watchgod==0.8.2
websockets==10.3
whitenoise==5.3.0
wsproto==0.15.0
xkit==0.0.0
yarl==1.7.2
youtube-video-upload==0.2.0
